See if non-default output device can be used
Trying to reduce midi-event->sound delays.

TODO To reduce delays try to pick other devices for output
(pick a device from cpal and make a rodio stream from it).
Experimented with sine wave source and there are similar delays after
the midi event as with a plugin-generated sound. So delayz are likely
at the output (buffering or the device).