Rename Lane -> Track
Previously have chosen a different name to not have name conflicts with
midly, but "Lane" seems confusing as we do not support multi-track
compositions, or one may think that "lane" refers to e.g. a pitch on the
stave.

Will have midly types always prefixed with `midly::` from now on.