Partial fix for note editing
Have to ensure the right order of events after edits.

The assertions did not trigger every time probably because current
display procedure draws all events are displayed not only visible ones.
So this hides the fact that new events were always added to the end of
the track event sequence independently of their time position.